requirements.txt
asgiref==3.5.1
Django==3.2
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary>=2.8
psycopg2
pytest==6.2.3
pytest-django==4.2.0
djangorestframework==3.13.1
django-rest-swagger==2.2.0
drf-yasg
graphene-django
pillow